requirements.txt
flask~=3.0.1
openai~=1.8.0
python-dotenv~=1.0.0
alembic~=1.13.1
SQLAlchemy~=2.0.25